pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "fincast"
version = "0.1.0"
description = "FinCast Financial Time series foundation model."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
dependencies = [
    "absl-py==2.2.2",
    "annotated-types==0.7.0",
    "beartype==0.20.2",
    "bottleneck==1.4.2",
    "brotli==1.1.0",
    "certifi==2025.1.31",
    "click==8.1.8",
    "colt5-attention==0.11.1",
    "einops==0.8.1",
    "einx==0.3.0",
    "huggingface-hub==0.30.1",
    "jax==0.5.3",
    "matplotlib==3.10.1",
    "numpy==2.1.3",
    "pandas==2.2.3",
    "peft==0.17.1",
    "pillow==11.1.0",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "tensorflow==2.19.0",
    "torch==2.5.0",
    "utilsforecast==0.2.12",
    "wandb==0.19.9",
    "yfinance==0.2.55",
]

[project.urls]
Repository = "https://github.com/vincent05r/FinCast-fts"

[tool.setuptools.packages.find]
where = ["src"]
